requirements.txt
# Updated: 2026-04-27 - Synchronous project (no Celery/Redis/Channels)
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
astroid==4.0.4
attrs==26.1.0
bcrypt==5.0.0
brotli==1.2.0
cbor2==5.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
constantly==23.10.4
cryptography==46.0.7
cssselect2==0.9.0
dill==0.4.1
distro==1.9.0
Django==5.2.13
django-cors-headers==4.9.0
django-ninja==1.6.2
et_xmlfile==2.0.0
eval_type_backport==0.3.1
filelock==3.29.0
fonttools==4.62.1
freetype-py==2.5.1
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.13
ImageIO==2.37.3
importlib_metadata==8.7.1
isort==8.0.1
Jinja2==3.1.6
jsonpath-python==1.1.5
lazy-loader==0.5
lxml==6.1.0
MarkupSafe==3.0.3
mccabe==0.7.0
mistralai==2.4.2
mpmath==1.3.0
mysqlclient==2.2.8
networkx==3.6.1
numpy==2.4.4
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
oscrypto==1.3.0
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
proto-plus==1.27.2
protobuf==5.29.6
pycairo==1.29.0
pyclipper==1.4.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydyf==0.12.1
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
PyJWT==2.12.1
pylint==4.0.5
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pypdf==6.10.2
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-magic==0.4.27
PyYAML==6.0.3
reportlab==4.4.10
requests==2.33.1
rlPyCairo==0.4.0
scikit-image==0.26.0
scipy==1.17.1
segno==1.6.6
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
svglib==1.6.0
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.4.11
tinycss2==1.5.1
tinyhtml5==2.1.0
tomlkit==0.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
uritools==6.0.2
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
weasyprint==60.1
whitenoise==6.12.0
xhtml2pdf==0.2.17
zipp==3.23.1
zopfli==0.4.1
dj_database_url
psycopg2-binary

# ═══ Additional dependencies for new features ═══
# File upload & document management
python-multipart>=0.0.6

# Calendar & timezone
pytz>=2024.1

# File type detection
filetype>=1.2.0

# Video processing (optional)
moviepy>=1.0.3
